# server/swagger.yaml
openapi: 3.0.0
info:
  title: Business Nexus API
  version: 1.0.0
  description: API documentation for the Business Nexus platform bridging Entrepreneur & Investor connections.
servers:
  - url: /
    description: Current deployment environment
paths:
  /api/auth/login:
    post:
      summary: User Login
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, example: "sarah@techwave.io" }
                password: { type: string, example: "password123" }
                role: { type: string, enum: ['entrepreneur', 'investor'] }
      responses:
        '200':
          description: Login successful (returns JWT token or triggers 2FA)
  /api/auth/register:
    post:
      summary: User Registration
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
                role: { type: string, enum: ['entrepreneur', 'investor'] }
      responses:
        '201':
          description: User Created successfully
  /api/payments/transfer:
    post:
      summary: Transfer funds between accounts securely
      tags: [Payments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetUserId: { type: string }
                amount: { type: number }
                description: { type: string }
      responses:
        '200':
          description: Transfer executed atomically
  /api/meetings:
    get:
      summary: List all user meetings
      tags: [Meetings]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
    post:
      summary: Schedule a new meeting
      tags: [Meetings]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                entrepreneurId: { type: string }
                investorId: { type: string }
                date: { type: string }
                startTime: { type: string }
                durationMinutes: { type: number }
      responses:
        '201':
          description: Meeting scheduled avoiding conflicts
  /api/collaborations:
    post:
      summary: Request collaboration
      tags: [Collaborations]
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Request sent (Investor role protected)
        '403':
          description: Forbidden due to missing Investor role authorizations.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT